Move communication and annotations to AddressPage tabs
