+++ src/app/pages/AddressPage/Communication/index.tsx
@@ -1,12 +1,10 @@
 import { Row } from "@amsterdam/design-system-react"
-import { Annotation } from "./Annotation/Annotation"
 import { CreateCommunicationNote } from "./actions/CreateCommunicationNote"
 import { CommunicationNotesTable } from "./CommunicationNotesTable/CommunicationNotesTable"
 
 const Communication: React.FC = () => (
   <>
-    <Annotation />
-    <Row align="end">
+    <Row align="end" style={{ marginBottom: "1rem" }}>
       <CreateCommunicationNote />
     </Row>
     <CommunicationNotesTable />
