--- src/app/pages/AddressPage/Section.tsx
@@ -1,9 +1,0 @@
-type SectionProps = {
-  children: React.ReactNode
-}
-
-export const Section: React.FC<SectionProps> = ({ children }) => (
-  <div style={{ marginTop: "2rem" }}>{children}</div>
-)
-
-export default Section
